moved logic out of presentational components to parent 'pages'
