+++ src/main/front/components/Dashboard/Account/Account.js
@@ -1,83 +1,43 @@
-import React, { Component } from 'react';
+import React from 'react';
+import { array } from 'prop-types';
 
+/* Styles */
 import './Account.less';
 
-class Account extends Component {
-  constructor(props) {
-    super(props);
-  }
+const Account = props => (
+	<div id="dashboard-account">
+		<div id="account-user-info">
+			<p className="account-name"> Account Name </p>
+			<p className="account-email"> Email </p>
+			<p className="account-about"> Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. </p>
+		</div>
 
-	getArticles = () => {
-		let articles = [];
-
-		for (let j = 0; j < 4; j++) {
-			articles.push(
-				<div className="article" key={j}>
-					<div className="dashboard-img article-img">
-						<img src="https://picsum.photos/450/285" />
-					</div>
-
-					<p className="article-title"> Lorem ipsum dolor sit amet, consectetur adipiscing elit </p>
-				</div>
-			)
-		}
-
-		return articles;
-	}
-
-	getFollowers = () => {
-		let followers = [];
-
-		for (let j = 0; j < 5; j++) {
-			followers.push(
-				<div className="following" key={j}>
-					<div className="dashboard-img following-img">
-						<img src="https://i.pravatar.cc/120" />
-					</div>
-
-					<p className="following-name"> John Doe </p>
-				</div>
-			)
-		}
-
-		return followers;
-	}
-
-
-  render() {
-    return (
-		<div id="dashboard-account">
-
-			<div id="account-user-info">
-				<p className="account-name"> Account Name </p>
-				<p className="account-email"> Email </p>
-				<p className="account-about"> Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. </p>
+		<div id="account-article-saved" className="account-slider">
+			<h4> Saved Articles </h4>
+			<div className="inner">
+				{props.articles}
 			</div>
+		</div>
 
-			<div id="account-article-saved" className="account-slider">
-				<h4> Saved Articles </h4>
-				<div className="inner">
-					{this.getArticles()}
-				</div>
+		<div id="account-following" className="account-slider">
+			<h4> Following </h4>
+			<div className="inner">
+				{props.following}
 			</div>
+		</div>
 
-			<div id="account-following" className="account-slider">
-				<h4> Following </h4>
-				<div className="inner">
-					{this.getFollowers()}
-				</div>
-			</div>
+		<div id="account-article-topics">
+			<p className="topic active"> All Articles (17) </p>
+			<p className="topic"> Topic A (10) </p>
+			<p className="topic"> Topic B (3) </p>
+			<p className="topic"> Topic C (4) </p>
+		</div>
+	</div>
+);
 
-			<div id="account-article-topics">
-				<p className="topic active"> All Articles (17) </p>
-				<p className="topic"> Topic A (10) </p>
-				<p className="topic"> Topic B (3) </p>
-				<p className="topic"> Topic C (4) </p>
-			</div>			
-
-    	</div>
-    )
-  }
-}
+Account.propTypes = {
+	articles: array.isRequired,
+	following: array.isRequired,
+};
 
 export default Account;
